+++ core/src/main/java/net/akazukin/library/LibraryPlugin.java
@@ -15,13 +15,13 @@ import java.util.logging.LogRecord;
 import java.util.logging.Logger;
 import net.akazukin.library.command.Command;
 import net.akazukin.library.command.LibraryCommandManager;
-import net.akazukin.library.compat.grimac.GrimACEvents;
 import net.akazukin.library.compat.minecraft.Compat;
 import net.akazukin.library.compat.minecraft.CompatManager;
 import net.akazukin.library.doma.LibrarySQLConfig;
 import net.akazukin.library.doma.dao.MUserDaoImpl;
 import net.akazukin.library.doma.dao.MUserProfileDaoImpl;
 import net.akazukin.library.event.Events;
+import net.akazukin.library.event.GrimACEvents;
 import net.akazukin.library.event.LibraryEventManager;
 import net.akazukin.library.event.events.ServerTickEvent;
 import net.akazukin.library.gui.GuiManager;
